+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ##### Changed
 * 💥 **\[core\]** Add a parameter `onStateUpdateComplete` to `Component.scheduleStateUpdate(...)` that is
-  invoked by the react runtime after the state update has completed.
+  invoked by the react runtime after the state update has completed. Add additional overloaded variants of
+  `Component.scheduleStateUpdate(...)` where the call back is not present. This had the impact of changing
+  the variant without the `onStateUpdateComplete` parameter to be a final method invocation.
 
 ### [v0.23](https://github.com/react4j/react4j/tree/v0.23) (2017-12-10)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.22...v0.23)